+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 # not yet released, version 0.2.0
 
 - Implemented support for `{lat, lng}` format, and corresponding function `toLatLng`.
+- Implemented `getLatitude` and `getLongitude`.
+- Implemented `angleAndDistanceTo`, `angleTo`, `distanceTo`.
 - Implemented conversion functions `degToRad`, `radToDeg`, `knotsToMeterPerSecond`, `meterPerSecondToKnots`, `knotsToKmPerHour`, and `kmPerHourToKnots`.
-
 
 # 2017-05-22, version 0.1.0